Fix SIGPIPE on PT server shutdown (fixes #11)
This is still a race condition. We ask the PT server to terminate
using aisc_nput(..., MAIN_SHUTDOWN) and subsequently call aisc_close()
which appears to write on the connection as well, causing the
SIGPIPE. Since the PT server still writes some messages to STDERR on
shutdown, we should be much faster than it though.
Fix SIGPIPE on PT server shutdown